fix: Se corrige tipo de dato de precio y peso de productos

+++ src/main/java/com/app/dao/ProductDao.java
@@ -18,7 +18,7 @@ public interface ProductDao extends CrudRepository<Product, Long> {
 	
 	@Modifying
 	@Query(value = "UPDATE Product p SET p.productName = ?1, p.reference = ?2, p.price = ?3, p.weight = ?4, p.stock = ?5, p.category = ?6 WHERE p.productId = ?7")
-	public void updateProduct(String productName, String reference, Float price, Float weight, int stock, Category category, Long productId);
+	public void updateProduct(String productName, String reference, int price, int weight, int stock, Category category, Long productId);
 
 	@Modifying
 	@Query(value = "UPDATE Product p SET p.stock = ?1 WHERE p.productId = ?2")
